+++ Team3_front/front/src/pages/MyPage/Mypage_artist.css
@@ -1,0 +1,186 @@
+/* General Reset */
+* {
+    margin: 0;
+    padding: 0;
+    box-sizing: border-box;
+  }
+  
+  /* Page Styling */
+  .my-page {
+    width: 100vw;
+    height: 100vh;
+    font-family: Arial, sans-serif;
+    color: #333;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  
+  .header {
+    background-color: #ddd;
+    width: 100%;
+    text-align: center;
+    padding: 20px 0;
+    font-size: 1.5em;
+    font-weight: bold;
+    flex: 0 0 8%; /* header가 화면의 약 8%를 차지 */
+  }
+  
+  /* Profile Section */
+  .profile-section {
+    display: flex;
+    width: 80%;
+    align-items: center;
+    flex: 0 0 30%; /* profile-section이 화면의 약 20% 차지 */
+  }
+  
+  .profile-image-container {
+    position: relative; /* edit-icon 배치를 위한 relative 설정 */
+    width: 200px;
+    height: 200px;
+  }
+  
+  .profile-image {
+    font-size: 100px; /* 아이콘 크기 */
+    color: #555; /* 아이콘 색상 */
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    height: 100%;
+    border-radius: 50%; /* 동그랗게 만들기 */
+    background-color: #f0f0f0; /* 배경색 */
+    overflow: hidden;
+    margin-left: 200px;
+  }
+  
+  .profile-image img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+  }
+  
+  /* Edit Icon */
+  .edit-icon {
+    position: absolute; /* profile-image 위에 겹치도록 설정 */
+    bottom: 30px; /* 프로필 이미지의 하단에 배치 */
+    right: -200px; /* 프로필 이미지의 오른쪽 끝에 배치 */
+    background-color: #ffffff;
+    color: #555;
+    font-size: 20px;
+    border-radius: 50%;
+    padding: 5px;
+    cursor: pointer;
+    box-shadow: 0 2px 5px rgba(0, 0, 0, 0.2);
+  }
+  
+  .edit-icon:hover {
+    background-color: #eeeeee;
+  }
+  
+  .profile-info p {
+    margin: 20px 0px;
+    padding-left: 50px;
+  }
+  
+  .profile-info {
+    margin-left: 500px;
+  }
+  
+  /* Interests Section */
+  .interests-section {
+    width: 90%;
+    border: 2px solid #e5a23b;
+    padding: 10px;
+    flex: 0 0 30%; /* 각 섹션이 화면의 약 30% 차지 */
+  }
+  
+  .interests-header {
+    display: flex; /* Flexbox로 정렬 */
+    justify-content: space-between; /* 좌우 정렬 */
+    align-items: center; /* 세로 중앙 정렬 */
+    margin-bottom: 10px;
+  }
+  
+  .interest-name {
+    font-size: 1.2em;
+    font-weight: bold;
+  }
+  
+  .edit-text {
+    font-size: 1em;
+    background-color: #e5a23b;
+    padding:5px 20px 5px 20px;
+    cursor: pointer;
+    text-decoration: none; /* 밑줄 제거 */
+  }
+  
+  
+  .interests-images {
+    display: flex;
+    gap: 10px;
+    justify-content: center;
+    overflow-x: auto; /* 가로 스크롤 활성화 */
+    padding-bottom: 10px;
+    white-space: nowrap; /* 항목이 줄바꿈되지 않고 한 줄에 표시 */
+  }
+  
+  /* Hover Effect for Interests */
+  .interest-image {
+    width: 150px;
+    height: 150px;
+    background-color: #ccc;
+    margin: 10px 50px;
+    display: inline-block; /* 가로로 정렬되도록 설정 */
+    transition: background-color 0.3s; /* 부드러운 전환 효과 */
+  }
+  
+  .interest-image:hover {
+    background-color: rgba(64, 64, 64, 0.7); /* 투명한 진한 회색 */
+  }
+  
+  /* Auction Section */
+  .auction-section {
+    width: 90%;
+    border-bottom: 2px solid #e5a23b;
+    border-left: 2px solid #e5a23b;
+    border-right: 2px solid #e5a23b;
+    padding: 10px;
+    margin-bottom: 10px;
+    flex: 0 0 30%; /* 각 섹션이 화면의 약 30% 차지 */
+  }
+  
+  .auction-name,
+  .interest-name {
+    margin: 10px 10px;
+  }
+  
+  .auction-item-list {
+    max-height: 200px; /* 높이 제한 */
+    overflow-y: auto; /* 세로 스크롤 활성화 */
+  }
+  
+  .auction-item {
+    display: flex;
+    justify-content: space-between;
+    padding: 10px;
+    background-color: #eee;
+    margin-bottom: 10px;
+    margin: 20px 50px;
+    transition: background-color 0.3s; /* 부드러운 전환 효과 */
+    cursor: pointer; /* 커서 표시 */
+  }
+  
+  .auction-item:hover {
+    background-color: rgba(64, 64, 64, 0.7); /* 투명한 진한 회색 */
+  }
+  
+  .auction-item p,
+  .auction-item span {
+    margin: 0;
+  }
+  
+  .no-auctions
+  {
+    text-align: center;
+  }
